+++ Pacmio/Source/Modules/BarTable/Analysis/Indicators/Bollinger.cs
@@ -131,7 +131,7 @@ namespace Pacmio
 
         public string AreaName { get; private set; } = MainArea.DefaultName;
 
-        public int AreaRatio { get; set; } = 8;
+        public float AreaRatio { get; set; } = 8;
 
         public void ConfigChart(BarChart bc)
         {
