+++ Pacmio/Source/Modules/BarTable/Analysis/Indicators/Chanderlier.cs
@@ -113,7 +113,7 @@ namespace Pacmio
 
         public string AreaName { get; private set; } = MainArea.DefaultName;
 
-        public int AreaRatio { get; set; } = 8;
+        public float AreaRatio { get; set; } = 8;
 
         public int AreaOrder { get; set; } = 0;
